Refactor: 컴포넌트 분리 및 코드 리팩토링
