+++ src/components/domain/stays/index.ts
@@ -2,3 +2,5 @@ export { default as StayActionBar } from './StayActionBar';
 export { default as StayDescription } from './StayDescription';
 export { default as StayHeader } from './StayHeader';
 export { default as StayInfoChips } from './StayInfoChips';
+export { default as CityActionBar } from './CityActionBar';
+export { default as CountrysideActionBar } from './CountrysideActionBar';
